backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
requests
python-dotenv
motor
pymongo[srv]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
